requirements.txt
beautifulsoup4==4.12.3
blinker==1.7.0
Brlapi==0.8.5
Brotli==1.1.0
certifi==2023.11.17
chardet==5.2.0
click==8.1.6
colorama==0.4.6
command-not-found==0.3
configobj==5.0.8
cryptography==41.0.7
cupshelpers==1.0
dbus-python==1.3.2
defer==1.0.6
deprecation==2.0.7
dill==0.3.8
distro==1.9.0
eyed3==0.9.7
filelock==3.13.1
filetype==1.2.0
html2text==2024.2.26
httplib2==0.20.4
idna==3.6
ifaddr==0.2.0
IMDbPY==2021.4.18
launchpadlib==1.11.0
lazr.restfulclient==0.14.6
lazr.uri==1.0.6
libevdev==0.5
louis==3.29.0
lxml==5.2.1
Mako==1.3.2.dev0
Markdown==3.5.2
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
multiprocess==0.70.16
mutagen==1.46.0
nemo-emblems==6.4.0
netaddr==0.8.0
netifaces==0.11.0
oauthlib==3.2.2
onboard==1.4.1
packaging==24.0
PAM==0.4.2
pexpect==4.9.0
pillow==10.2.0
psutil==5.9.8
ptyprocess==0.7.0
pyasyncore==1.0.2
pycairo==1.25.1
pycryptodomex==3.20.0
pycups==2.0.1
pycurl==7.45.3
pyelftools==0.30
Pygments==2.17.2
PyGObject==3.48.2
PyICU==2.12
pyinotify==0.9.6
PyJWT==2.7.0
PyNaCl==1.5.0
pyparsing==3.1.1
pyparted==3.12.0
pypng==0.20231004.0
python-apt==2.7.7+ubuntu5
python-debian==0.1.49+ubuntu2
python-gnupg==0.5.2
python-magic==0.4.27
python-xlib==0.33
pytz==2024.1
pyudev==0.24.0
pyxdg==0.28
# Dependências do projeto
telethon==1.43.0
PyYAML==6.0.1
python-dotenv==1.0.0
qrcode==7.4.2
repolib==2.2.1
requests==2.31.0
requests-file==1.5.1
rich==13.7.1
setproctitle==1.3.3
setuptools==68.1.2
six==1.16.0
soupsieve==2.5
ssh-import-id==5.11
systemd-python==235
tinycss2==1.2.1
tldextract==3.1.2
typing_extensions==4.10.0
ubuntu-drivers-common==0.0.0
ufw==0.36.2
Unidecode==1.3.8
urllib3==2.0.7
wadllib==1.3.6
webencodings==0.5.1
websockets==10.4
wheel==0.42.0
xdg==5
xkit==0.0.0
xlrd==2.0.1
yt-dlp==2024.4.9
